pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "influxdb-client>=1.48.0",
    "librouteros>=3.4.1",
    "psutil>=7.0.0",
    "pyjwt>=2.10.1",
    "pysnmp>=7.1.16",
    "pyyaml>=6.0.2",
]
